Added a data model and database client to RankingService to manage its own data for rankings.
Added a field Timestamp to RecipePost objects in the RecipeCatalogue database. This stores the creation
time of the recipe post, and will be used by RankingService to determine which recipes to rank based
on their creation time (rather than, e.g. the time in which RankingService received the creation event).